refactor: split filter ui into files
